requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2==2.8.6
pillow==9.4.0
django-filter==22.1
# celery[redis]==5.2.7
# flower==1.2.0
# celery_singleton==0.3.1
# django_redis==5.2.0
# django_cachalot==2.5.2